+++ CHANGELOG.md
@@ -5,6 +5,7 @@
   * Added `Skogsra.Type` behaviour for defining custom types.
   * Application configuration values are now casted as well to the defined
     type in the `Skogsra` environment variable definition.
+  * Added YAML configuration provider.
 
 # Changelog for 1.3.0